Separated FastAPI-functionality into it's own folder. Updated server running to use python scripts. Updated README to reflect running changes
