+++ server/Dockerfile
@@ -5,4 +5,7 @@ RUN pip install -r requirements.txt
 
 COPY . .
 EXPOSE 8080
-CMD ["uvicorn", "src.main:application", "--host", "0.0.0.0", "--port", "8080"]
+ENV HOST=0.0.0.0
+ENV PORT="8080"
+ENV RELOAD="False"
+CMD ["python", "src/main.py"]
