+++ server/src/api/routes.py
@@ -1,0 +1,6 @@
+from api.application import application
+
+
+@application.get("/")
+async def root():
+    return {"message": "Hello Worlds"}
